requirements.txt
oslo.config>=1.2.0
oslo.i18n>=0.1.0
oslo.serialization>=1.0.0 # Apache-2.0
oslo.utils>=1.0.0 # Apache-2.0
stevedore>=0.12

python-keystoneclient>=0.11.1
python-cinderclient>=1.0.6

# Required by openstack.common libraries
eventlet>=0.13.0
iso8601>=0.1.8
six>=1.4.1

# py2.6 compat
ordereddict
